+++ src/features/user/user.composition.ts
@@ -1,10 +1,10 @@
 import { UserFeature } from './user';
 import { UserApiService } from 'services/user';
-import { FirebaseUserApiRepository } from 'repositories/user/api/api';
+import { UserFirebaseApiRepository } from 'repositories/user/api/api';
 import { UserStoreApiService } from 'services/user/store-api';
 import { ReduxUserStoreApiRepository } from 'repositories/user/store-api/store-api';
 import { type FirebaseApi } from 'repositories/firebase';
-import { type ReduxStoreApi } from 'repositories/store';
+import { type ReduxStoreApi } from 'repositories/redux';
 
 interface Dependencies {
     api: FirebaseApi
@@ -12,7 +12,7 @@ interface Dependencies {
 };
 
 export const createUserComposition = (deps: Dependencies): UserFeature => {
-    const apiRepository = new FirebaseUserApiRepository(deps.api);
+    const apiRepository = new UserFirebaseApiRepository(deps.api);
     const apiService = new UserApiService(apiRepository);
 
     const storeApiRepository = new ReduxUserStoreApiRepository(deps.storeApi);
