+++ src/repositories/post-card-list/store-api.ts
@@ -1,6 +1,6 @@
+import { ReduxApiRepository } from 'repositories/redux';
 import { postCardListSlice } from './store-slice';
 import { type PostCard } from 'models/post-card-list';
-import { ReduxApiRepository } from 'repositories/types';
 
 export class PostStoreApiRepository extends ReduxApiRepository {
     public async readModel (count: number): Promise<PostCard[]> {
