+++ src/repositories/types.ts
@@ -1,4 +1,4 @@
-import { type ReduxStoreApi } from 'repositories/store';
+import { type ReduxStoreApi } from 'repositories/redux';
 import { type FirebaseApi } from './firebase';
 
 export type Api = FirebaseApi | ReduxStoreApi;
@@ -8,30 +8,5 @@ export abstract class ApiRepository {
 
     constructor (api: Api) {
         this.api = api;
-    }
-}
-
-export class ApiError extends Error {
-    type: string;
-
-    constructor (e: Error) {
-        super(e.message);
-        this.type = e.message;
-    }
-}
-
-export abstract class ReduxApiRepository extends ApiRepository {
-    readonly api: ReduxStoreApi;
-
-    constructor (api: ReduxStoreApi) {
-        super(api);
-    }
-}
-
-export abstract class FirebaseApiRepository extends ApiRepository {
-    readonly api: FirebaseApi;
-
-    constructor (api: FirebaseApi) {
-        super(api);
     }
 }
