+++ src/repositories/user/index.ts
@@ -1,3 +1,3 @@
 export type { UserApiRepository } from './api/types';
-export type { UserStoreApiRepository } from 'repositories/user/store-api/store-api';
+export type { UserStoreApiRepository } from 'repositories/user/store-api/types';
 export { UserAuthError } from './api/errors';
